CE-1857 Abstract API controller
