+++ extensions/wikia/Flags/controllers/FlagsController.class.php
@@ -9,239 +9,33 @@
  * @license http://www.gnu.org/copyleft/gpl.html GNU General Public License 2.0 or later
  */
 
-use Flags\FlagsCache;
+use Flags\Views\FlagView;
 use Flags\FlagsHelper;
-use Flags\Models\Flag;
-use Flags\Models\FlagType;
 
-class FlagsController extends WikiaApiController {
+class FlagsController extends WikiaController {
 	private
-		$cache,
-		$params,
-		$status = false;
+		$params;
 
 	/**
-	 * Generates html contents for Flags modal for editing flags
-	 * @TODO Move modal away from API to client controller
+	 * A wrapper for a request to the FlagsApiController for flags for the given page
+	 * that returns a wikitext string with calls to templates of the flags.
+	 * @return null|string
 	 */
-	public function editForm() {
-		global $wgUser;
+	public function getFlagsForPageWikitext( $pageId ) {
+		$flags = $this->requestGetFlagsForPage( $pageId );
 
-		$pageId = $this->request->getVal( 'pageId' );
-		if ( empty( $pageId ) ) {
-			$this->response->setException( new \Exception( 'Required param pageId not provided' ) );
-			return true;
+		if ( !empty( $flags ) ) {
+			$flagsWikitext = '';
+
+			$flagView = new FlagView();
+			foreach ( $flags as $flagId => $flag ) {
+				$flagsWikitext .= $flagView->createWikitextCall( $flag['flag_view'], $flag['params'] );
+			}
+
+			return $flagsWikitext;
 		}
 
-		$flagsWikiaResponse = $this->app->sendRequest( 'FlagsController', 'getFlagsForPageForEdit', [
-			'pageId' => $pageId
-		] );
-
-		$flags = $flagsWikiaResponse->getData();
-
-		$this->setVal( 'editToken', $wgUser->getEditToken() );
-		$this->setVal( 'flags', $flags );
-		$this->setVal( 'formSubmitUrl', $this->getLocalUrl('postFlagsEditForm') );
-		$this->setVal( 'inputNamePrefix', FlagsHelper::FLAGS_INPUT_NAME_PREFIX );
-		$this->setVal( 'inputNameCheckbox', FlagsHelper::FLAGS_INPUT_NAME_CHECKBOX );
-		$this->setVal( 'pageId', $pageId );
-
-		$this->response->setFormat('html');
-	}
-
-	/**
-	 * Returns a singleton instance of FlagsCache
-	 * @return FlagsCache
-	 */
-	private function getCache() {
-		if ( !isset( $this->cache ) ) {
-			$this->cache = new FlagsCache();
-		}
-
-		return $this->cache;
-	}
-
-	/**
-	 * Assigns a request's parameters to the object's property
-	 * and sets a wikiId if it hasn't been specified as one
-	 * of the parameters.
-	 */
-	private function getRequestParams() {
-		$this->params = $this->request->getParams();
-		if ( !isset( $this->params['wikiId'] ) ) {
-			$this->params['wikiId'] = $this->wg->CityId;
-		}
-	}
-
-	/**
-	 * To prevent CSRF attacks it checks if a request is a POST one
-	 * and if a sent token matches the user's one.
-	 * Calls getRequestParams if the request is valid.
-	 * @return bool
-	 */
-	private function processRequest() {
-		if ( !$this->request->wasPosted()
-			|| !$this->wg->User->matchEditToken( $this->getVal( 'token' ) )
-		) {
-			$this->response->setException( new \Exception( 'Invalid request' ) );
-			return false;
-		}
-
-		$this->getRequestParams();
-		return true;
-	}
-
-	/**
-	 * Article level API
-	 */
-
-	/**
-	 * Retrieves all data for all flag types available on the given wikia
-	 * with an intent of rendering a modal with an edit form.
-	 * It returns the types with instances on the page first and then all
-	 * other types sorted alphabetically.
-	 * To retrieve only types with instances on the given page:
-	 * @see getFlagsForPage()
-	 *
-	 * @requestParam int wikiId
-	 * @requestParam int pageId
-	 * @response Array A list of flags with flag_type_id values as indexes.
-	 *  One item contains the following fields:
-	 * @if The page has an instance of the flag type
-	 *	 	int flag_id
-	 *
-	 * 		int flag_type_id
-	 * 		int wiki_id
-	 *		int page_id
-	 * 		int flag_group
-	 * 		string flag_name
-	 * 		string flag_view A name of a template of the flag
-	 * 		string flag_view_url A full URL of the template
-	 * 		int flag_targeting
-	 * 		string|null flag_params_names
-	 *
-	 * 	@if flag_params_names is not empty
-	 * 		params = [
-	 * 			param_name => param_value
-	 *		]
-	 */
-	public function getFlagsForPageForEdit() {
-		$this->getRequestParams();
-
-		if ( !isset( $this->params['pageId'] ) ) {
-			return null;
-		}
-
-		$allFlagTypes = $this->getAllFlagTypes( $this->params['wikiId'], $this->params['pageId'] );
-
-		$this->setResponseData( $allFlagTypes );
-	}
-
-	/**
-	 * A method that combines data from two models - Flag and FlagTypes.
-	 * The Flag model gives it data on instances of flags assigned to the given page.
-	 * The FlagType model provides data on every type of a flag available on the wikia.
-	 * The union operator used on arrays lets us safely merge these two arrays.
-	 * In case of an index duplication it preserves an item from the left array and adds a new
-	 * item if an index is not set.
-	 *
-	 * @param $wikiId
-	 * @param $pageId
-	 * @return Array An array of all types of flags, with and without instances
-	 */
-	private function getAllFlagTypes( $wikiId, $pageId ) {
-		/**
-		 * 1. Get flags assigned to the page
-		 */
-		$flagsForPage = $this->getFlagsForPageRawData( $wikiId, $pageId );
-
-		/**
-		 * 2. Get all flag types for a wikia
-		 */
-		$flagTypesForWikia = $this->getFlagTypesForWikiaRawData( $wikiId );
-
-		/**
-		 * 3. Return the united arrays if it's possible to merge them (both arrays use
-		 * flag_type_id values as indexes)
-		 */
-		return $flagsForPage + $flagTypesForWikia;
-	}
-
-	/**
-	 * Retrieves all data for flags assigned to the given page
-	 * with an intent of rendering them. To get all types of flags:
-	 * @see getFlagsForPageForEdit()
-	 *
-	 * @requestParam int wikiId
-	 * @requestParam int pageId
-	 * @response Array A list of flags with flag_type_id values as indexes.
-	 *  One item contains the following fields:
-	 *	 	int flag_id
-	 * 		int flag_type_id
-	 * 		int wiki_id
-	 *		int page_id
-	 * 		int flag_group
-	 * 		string flag_name
-	 * 		string flag_view
-	 * 		int flag_targeting
-	 * 		string|null flag_params_names
-	 *
-	 * 	@if flag_params_names is not empty
-	 * 		params = [
-	 * 			param_name => param_value
-	 *		]
-	 */
-	public function getFlagsForPage() {
-		$this->getRequestParams();
-
-		if ( !isset( $this->params['pageId'] ) ) {
-			return null;
-		}
-
-		$flagsForPage = $this->getFlagsForPageRawData( $this->params['wikiId'], $this->params['pageId'] );
-
-		$this->setResponseData( $flagsForPage );
-	}
-
-	/**
-	 * Tries to get the data on Flags instances for the given page from cache.
-	 * If it is not cached it gets it from the database and caches it.
-	 * @param $wikiId
-	 * @param $pageId
-	 * @return bool|mixed
-	 */
-	private function getFlagsForPageRawData( $wikiId, $pageId ) {
-		$flagsCache = $this->getCache();
-		$flagsForPage = $flagsCache->getFlagsForPage( $pageId );
-
-		if ( !$flagsForPage ) {
-			$flagModel = new Flag();
-			$flagsForPage = $flagModel->getFlagsForPage( $wikiId, $pageId );
-
-			$flagsCache->setFlagsForPage( $pageId, $flagsForPage );
-		}
-
-		return $flagsForPage;
-	}
-
-	/**
-	 * Tries to get the data on types of flags available for the given wikia from cache.
-	 * If it is not cached it gets it from the database and caches it.
-	 * @param $wikiId
-	 * @return bool|mixed
-	 */
-	private function getFlagTypesForWikiaRawData( $wikiId ) {
-		$flagsCache = $this->getCache();
-
-		$flagTypesForWikia = $flagsCache->getFlagTypesForWikia();
-		if ( !$flagTypesForWikia ) {
-			$flagTypeModel = new FlagType();
-			$flagTypesForWikia = $flagTypeModel->getFlagTypesForWikia( $wikiId );
-
-			$flagsCache->setFlagTypesForWikia( $flagTypesForWikia );
-		}
-
-		return $flagTypesForWikia;
+		return null;
 	}
 
 	/**
@@ -250,8 +44,8 @@ class FlagsController extends WikiaApiController {
 	 * matches an edit token for $wgUser.
 	 *
 	 * The request should include:
-	 * @requestParam int wikiId (if not provided a $wgCityId value is used)
-	 * @requestParam int pageId
+	 * @requestParam int wiki_id (if not provided a $wgCityId value is used)
+	 * @requestParam int page_id
 	 * @requestParam string token
 	 *
 	 * Input fields of the form should have a prefix `editFlags:flag_type_id:`
@@ -261,13 +55,11 @@ class FlagsController extends WikiaApiController {
 	 */
 	public function postFlagsEditForm() {
 		$this->processRequest();
-
 		if ( !isset( $this->params['pageId'] ) ) {
 			$this->response->setException( new \Exception( 'Required param pageId not provided' ) );
 			return true;
 		}
-		$wikiId = $this->params['wikiId'];
-		$pageId = $this->params['pageId'];
+		$pageId = $this->params['page_id'];
 
 		$title = Title::newFromID( $pageId );
 		if ( $title === null ) {
@@ -278,12 +70,12 @@ class FlagsController extends WikiaApiController {
 		/**
 		 * Get the current status to compare
 		 */
-		$currentFlags = $this->getAllFlagTypes( $wikiId, $pageId );
+		$currentFlags = $this->requestGetFlagsForPageForEdit( $pageId );
 
 		$helper = new FlagsHelper();
 		$flagsToChange = $helper->compareDataAndGetFlagsToChange( $currentFlags, $this->params );
 		if ( !empty( $flagsToChange ) ) {
-			$this->performActionsUsingPostedData( $flagsToChange );
+			$this->sendRequestsUsingPostedData( $flagsToChange );
 
 			/**
 			 * Purge cache values for the page
@@ -314,15 +106,14 @@ class FlagsController extends WikiaApiController {
 	 * 3. `toRemove` an array with IDs of flags to remove
 	 * @param Array $flagsToChange an array with three possible nested arrays:
 	 */
-	private function performActionsUsingPostedData( Array $flagsToChange ) {
-		$flagModel = new Flag();
+	private function sendRequestsUsingPostedData( Array $flagsToChange ) {
 		/**
 		 * Add flags
 		 */
 		if ( !empty( $flagsToChange['toAdd'] ) ) {
 			$flagsToAdd = [
-				'wikiId' => $this->params['wikiId'],
-				'pageId' => $this->params['pageId'],
+				'wiki_id' => $this->params['wiki_id'],
+				'page_id' => $this->params['page_id'],
 				'flags' => $flagsToChange['toAdd'],
 			];
 
@@ -352,98 +143,41 @@ class FlagsController extends WikiaApiController {
 	}
 
 	/**
-	 * Adds flags to the given page. It accepts only POST requests
-	 * with a valid User edit token.
-	 *
-	 * Required parameters:
-	 * @requestParam int wikiId
-	 * @requestParam int pageId
-	 * @requestParam array flags
-	 * @requestParam int flags['flag_type_id'] An ID of a flag type
-	 *
-	 * Optional parameters:
-	 * @requestParam array flags['params'] An array of params structured like:
-	 * [
-	 * 	'paramName1' => 'paramValue1',
-	 * 	'paramName2' => 'paramValue2',
-	 * ]
+	 * Sends a request to the FlagsApiController to get data on flags for the given page.
+	 * @param int $pageId
+	 * @return array
 	 */
-	public function addFlagsToPage() {
-		$this->processRequest();
-		$flagModel = new Flag();
-
-		if ( $flagModel->verifyParamsForAdd( $this->params ) ) {
-			$this->status = $flagModel->addFlagsToPage( $this->params );
-		}
-
-		$this->setVal( 'status', $this->status );
+	private function requestGetFlagsForPage( $pageId ) {
+		return $this->sendRequest( 'FlagsApiController',
+			'getFlagsForPage',
+			[
+				'page_id' => $pageId,
+			]
+		)->getData();
 	}
 
 	/**
-	 * Removes flags from the given page. It accepts only POST requests
-	 * with a valid User edit token.
-	 *
-	 * Required parameters:
-	 * @requestParam array flagsIds An array of IDs of flags to remove
+	 * Sends a request to the FlagsApiController to get data on flag types
+	 * with and without instances to display in the edit form.
+	 * @param int $pageId
+	 * @return array
 	 */
-	public function removeFlagsFromPage() {
-		$this->processRequest();
-		$flagModel = new Flag();
-
-		if ( $flagModel->verifyParamsForRemove( $this->params ) ) {
-			$this->status = $flagModel->removeFlagsFromPage( $this->params );
-		}
-
-		$this->setVal( 'status', $this->status );
+	private function requestGetFlagsForPageForEdit( $pageId ) {
+		return $this->sendRequest( 'FlagsApiController',
+			'getFlagsForPageForEdit',
+			[
+				'page_id' => $pageId,
+			]
+		)->getData();
 	}
 
-	/**
-	 * Flag type level API
-	 */
-
-	/**
-	 * Adds a new type of flags.
-	 *
-	 * Required parameters:
-	 * @requestParam int wikiId
-	 * @requestParam int flagGroup One of the keys in flagGroups property of the FlagType model
-	 * @requestParam string flagName A name of the flag (not longer than 128 characters)
-	 * @requestParam string flagView A title of a template used for rendering the flag
-	 * @requestParam int flagTargeting A level of targeting: 0 -> readers, 1 -> contibutors, 2 -> admins
-	 *
-	 * Optional parameters:
-	 * @requestParam string flagParamsNames A JSON-encoded array of names of parameters
-	 * 		It's used for rendering inputs in the "Add a flag" form.
-	 */
-	public function addFlagType() {
-		$this->processRequest();
-		$flagTypeModel = new FlagType();
-
-		if ( $flagTypeModel->verifyParamsForAdd( $this->params ) ) {
-			$this->status = $flagTypeModel->addFlagType( $this->params );
-		}
-
-		$this->setVal( 'status', $this->status );
-	}
-
-	/**
-	 * Removes a type of flags.
-	 *
-	 * Required parameters:
-	 * @requestParam int flagTypeId
-	 *
-	 * IMPORTANT!
-	 * When using this method be aware that it removes ALL instances of this type
-	 * of flags with ALL of their parameters per the database's configuration.
-	 */
-	public function removeFlagType() {
-		$this->processRequest();
-		$flagTypeModel = new FlagType();
-
-		if ( $flagTypeModel->verifyParamsForRemove( $this->params ) ) {
-			$this->status = $flagTypeModel->removeFlagType( $this->params );
-		}
-
-		$this->setVal( 'status', $this->status );
+	private function requestAddFlagsToPage( $pageId, $flags ) {
+		return $this->sendRequest( 'FlagsApiController',
+			'addFlagsToPage',
+			[
+				'page_id' => $pageId,
+				'flags' => $flags,
+			]
+		)->getData();
 	}
 }
